refactor: Split realtime.py into multi-file module
- Analyzed 1478-line realtime.py and identified logical components
- Created modular structure with separate files for:
  - types.py: Protocol and type definitions
  - connection_management.py: Connection handling and lifecycle
  - event_handling.py: Event forwarding and callback management
  - subscriptions.py: User and market data subscription management
  - core.py: Main ProjectXRealtimeClient class
- Fixed linting errors with ruff
- Added assertions to handle mypy type checking for optional attributes
- Tested all imports work correctly
- Maintained backward compatibility with existing imports
