+++ src/project_x_py/realtime/__init__.py
@@ -1,0 +1,10 @@
+"""
+Real-time client module for ProjectX Gateway API WebSocket connections.
+
+This module provides the ProjectXRealtimeClient class for managing real-time
+connections to ProjectX SignalR hubs.
+"""
+
+from .core import ProjectXRealtimeClient
+
+__all__ = ["ProjectXRealtimeClient"]
